ファイル分けた、関数名変数名統一した
